RateLimiter tests: fix `coroutine 'coro' was never awaited` warning on pypy
There is a strange peculiarity with PyPy. The `coroutine 'coro' was
never awaited` warning is expected to be raised in this test, and
it is expected to be caught by `with warnings.catch_warnings(record=True)`.
This works fine on cpython.

On PyPy, however, the warning appears to be raised with a delay:

```
test/geocoders/pickpoint.py::TestPickPoint::test_reverse_language_parameter
  /home/travis/virtualenv/pypy3.6-7.1.1/site-packages/_pytest/_code/code.py:292: RuntimeWarning: coroutine 'test_sync_raises_for_awaitable.<locals>.g.<locals>.coro' was never awaited
```

Note the name of the test: it should have been `test/extra/rate_limiter.py`,
but actually it is some of the geocoders tests executed later.

It looks like the unawaited coroutines detection happens with a delay
on PyPy.

So the fix is to stop capturing warnings and return a task instead
of a coroutine. An unawaited task (fortunately) doesn't cause
the warning, so we can get away from this issue without actually
solving it.